Move resolvers to package
