+++ src/main/kotlin/com/freya02/bot/resolvers/LibraryTypeResolver.kt
@@ -1,5 +1,7 @@
-package com.freya02.bot.versioning
+package com.freya02.bot.resolvers
 
+import com.freya02.bot.utils.Utils.isBCGuild
+import com.freya02.bot.versioning.LibraryType
 import com.freya02.botcommands.api.BContext
 import com.freya02.botcommands.api.parameters.ParameterResolver
 import com.freya02.botcommands.api.parameters.SlashParameterResolver
@@ -10,14 +12,21 @@ import net.dv8tion.jda.api.interactions.commands.CommandInteractionPayload
 import net.dv8tion.jda.api.interactions.commands.OptionMapping
 import net.dv8tion.jda.api.interactions.commands.OptionType
 
-class ScriptTypeResolver : ParameterResolver<ScriptTypeResolver, ScriptType>(ScriptType::class), SlashParameterResolver<ScriptTypeResolver, ScriptType> {
+class LibraryTypeResolver : ParameterResolver<LibraryTypeResolver, LibraryType>(LibraryType::class), SlashParameterResolver<LibraryTypeResolver, LibraryType> {
     override val optionType: OptionType = OptionType.STRING
 
     override fun getPredefinedChoices(guild: Guild?): Collection<Command.Choice> {
-        return ScriptType.values()
-            .map { type ->
-                Command.Choice(type.name.lowercase().replaceFirstChar { c -> c.uppercaseChar() }, type.name)
-            }
+        return when {
+            guild.isBCGuild() -> listOf(
+                Command.Choice("BotCommands", LibraryType.BOT_COMMANDS.name),
+                Command.Choice("JDA 5", LibraryType.JDA5.name),
+                Command.Choice("JDA-KTX", LibraryType.JDA_KTX.name)
+            )
+            else -> listOf(
+                Command.Choice("JDA 5", LibraryType.JDA5.name),
+                Command.Choice("JDA-KTX", LibraryType.JDA_KTX.name)
+            )
+        }
     }
 
     override fun resolve(
@@ -25,5 +34,5 @@ class ScriptTypeResolver : ParameterResolver<ScriptTypeResolver, ScriptType>(Scr
         info: SlashCommandInfo,
         event: CommandInteractionPayload,
         optionMapping: OptionMapping
-    ): ScriptType = ScriptType.valueOf(optionMapping.asString)
+    ): LibraryType = LibraryType.valueOf(optionMapping.asString)
 }
