+++ src/main/kotlin/com/freya02/bot/resolvers/ScriptTypeResolver.kt
@@ -1,6 +1,6 @@
-package com.freya02.bot.versioning
+package com.freya02.bot.resolvers
 
-import com.freya02.bot.utils.Utils.isBCGuild
+import com.freya02.bot.versioning.ScriptType
 import com.freya02.botcommands.api.BContext
 import com.freya02.botcommands.api.parameters.ParameterResolver
 import com.freya02.botcommands.api.parameters.SlashParameterResolver
@@ -11,21 +11,14 @@ import net.dv8tion.jda.api.interactions.commands.CommandInteractionPayload
 import net.dv8tion.jda.api.interactions.commands.OptionMapping
 import net.dv8tion.jda.api.interactions.commands.OptionType
 
-class LibraryTypeResolver : ParameterResolver<LibraryTypeResolver, LibraryType>(LibraryType::class), SlashParameterResolver<LibraryTypeResolver, LibraryType> {
+class ScriptTypeResolver : ParameterResolver<ScriptTypeResolver, ScriptType>(ScriptType::class), SlashParameterResolver<ScriptTypeResolver, ScriptType> {
     override val optionType: OptionType = OptionType.STRING
 
     override fun getPredefinedChoices(guild: Guild?): Collection<Command.Choice> {
-        return when {
-            guild.isBCGuild() -> listOf(
-                Command.Choice("BotCommands", LibraryType.BOT_COMMANDS.name),
-                Command.Choice("JDA 5", LibraryType.JDA5.name),
-                Command.Choice("JDA-KTX", LibraryType.JDA_KTX.name)
-            )
-            else -> listOf(
-                Command.Choice("JDA 5", LibraryType.JDA5.name),
-                Command.Choice("JDA-KTX", LibraryType.JDA_KTX.name)
-            )
-        }
+        return ScriptType.values()
+            .map { type ->
+                Command.Choice(type.name.lowercase().replaceFirstChar { c -> c.uppercaseChar() }, type.name)
+            }
     }
 
     override fun resolve(
@@ -33,5 +26,5 @@ class LibraryTypeResolver : ParameterResolver<LibraryTypeResolver, LibraryType>(
         info: SlashCommandInfo,
         event: CommandInteractionPayload,
         optionMapping: OptionMapping
-    ): LibraryType = LibraryType.valueOf(optionMapping.asString)
+    ): ScriptType = ScriptType.valueOf(optionMapping.asString)
 }
